ci: Update flake8 config
Ignore "imported but unused" warnings for __init__.py files as it's a
common pattern to use __init__.py to export the public API of a package.

Ignore directories we don't want flake8 to traverse.

Ignore docstring errors in the host tests. Docstrings in test
classes are less important than docstrings in production code.