TargetBase: Fix unused and undefined variables
Previously we tried to print an undefined variable during exception
handling. We used a bare `except` statement which would suppress all
possible exceptions including SystemExit and KeyboardInterrupt, print an
"Unexpected error" message and forward the exception further up the
stack. This does not seem like desirable behaviour, when a user hits
CTRL-C they probably don't need to see an "Unexpected error" message in
the log. Instead we now catch any exceptions derived from `Exception`
(still not ideal) and print the error message from it.

In another case we defined a local variable for an exception but it
wasn't used, so remove it.